+++ demo/components/DemoUseTemporalFormat.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useState } from "react";
 import useTemporalFormat from '../../src/useTemporalFormat';
 import useTemporalDateTime from '../../src/useTemporalDateTime';
 import './DemoCard.css';